remove old graphg, add some input handling
