+++ src/main/java/Main.java
@@ -1,10 +1,8 @@
-import new_graphg.GraphG;
-import new_graphg.Node;
-import new_graphg.Tree;
+import graphg.*;
 
 public class Main {
-    public static void main(String[] args) {
-        new GraphG(5,2);
+    public static void main(String[] args) throws CantBePlanarException, NoSolutionException {
+        new GraphG().getGraphs(5,2);
 
 
     }
