--- src/main/java/new_graphg/GraphG.java
@@ -1,50 +1,0 @@
-package new_graphg;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class GraphG {
-
-
-    public GraphG(int nodes, int edges) {
-        Tree tree = new Tree();
-        List<Tree> trees = new ArrayList<>();
-        trees.add(tree);
-        trees = calculateTrees(trees, nodes - 2);
-        trees=calculateGraphs(trees,edges);
-        for (Tree t : trees) {
-            System.out.println("------------");
-            System.out.println(t.toDot());
-        }
-    }
-
-    private List<Tree> calculateTrees(List<Tree> trees, int nodesLeft) {
-        if (nodesLeft <= 0) return trees;
-        else {
-            List<Tree> newTrees = new ArrayList<>();
-            for (Tree t : trees)
-                for (Node n : t.getNodes()) {
-                    Tree newTree = t.copy();
-                    if (newTree.connectNewNode(n.getIndex()))
-                        newTrees.add(newTree);
-                }
-            return calculateTrees(newTrees, nodesLeft - 1);
-        }
-    }
-
-    private List<Tree> calculateGraphs(List<Tree> graphs, int edgesLeft) {
-        if (edgesLeft <= 0) return graphs;
-        else {
-            List<Tree> newGraphs = new ArrayList<>();
-            for (Tree g : graphs)
-                for (Node n1 : g.getNodes())
-                    for (Node n2 : g.getNodes()){
-                        Tree newGraph = g.copy();
-                        if (n1.getIndex() != n2.getIndex() && newGraph.connectNodes(n1.getIndex(), n2.getIndex())){
-                            newGraphs.add(newGraph);
-                        }
-                    }
-            return calculateGraphs(newGraphs, edgesLeft - 1);
-        }
-    }
-}
